pyproject.toml
[tool.poetry]
name = "ai"
version = "0.1.0"
description = ""
authors = ["steven. <sveshkini@berkeley.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = {extras = ["all"], version = "^0.0.33"}
python-dotenv = "^0.21.0"
openai = "^0.25.0"
gpt-index = "^0.1.0"
click = "^8.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
